DataServer/requirements.txt
sanic
asyncio
asyncpg
python-dotenv